+++ owwatcher/main.py
@@ -10,6 +10,7 @@ from .owwatcher_logger_configurer import OWWatcherLoggerConfigurer
 from .owwatcher import OWWatcher
 import signal
 import sys
+from .syslog_alerter import SyslogAlerter
 
 # Creating null loggers allows pytest test suite to run as logging is not
 # necessarily configured for each unit test run.
@@ -32,8 +33,13 @@ def main():
         logger_configurer = OWWatcherLoggerConfigurer(options)
         configure_logging(options, logger_configurer)
         fab = FileArchiverBuilder(_LOGGER, options.archive_path)
+        # TODO: Handle perms_mask default in Options instead of here
+        if options.perms_mask is None:
+            options.perms_mask = 0o002
+        syslog_alerter = SyslogAlerter(options.perms_mask, _LOGGER, _SYSLOG_LOGGER, is_snap=is_snap)
+        syslog_alerter.run()
         _OWWATCHER = OWWatcher(options.perms_mask, fab, _LOGGER,
-                               _SYSLOG_LOGGER,is_snap)
+                               syslog_alerter, is_snap=is_snap)
 
         register_signal_handlers()
     except Exception as ex: