+++ tests/test_owwatcher.py
@@ -18,12 +18,8 @@ class MockFileArchiverBuilder(file_archiver_builder.FileArchiverBuilder):
         return MagicMock()
 
 class OWWatcherTest(owwatcher.OWWatcher):
-    def __init__(self, monkeypatch, perms_mask, logger, syslog_logger, is_snap=False):
-        super().__init__(perms_mask, MockFileArchiverBuilder(), logger, syslog_logger, is_snap=False)
-        self.archive_queue_timeout_sec = .01
-
-        syslog_logger.warning = MagicMock()
-        syslog_logger.info = MagicMock()
+    def __init__(self, monkeypatch, perms_mask, logger, is_snap=False):
+        super().__init__(perms_mask, MockFileArchiverBuilder(), logger, MagicMock(), is_snap=is_snap)
 
     def _process_event(self, watch_dir, event):
         super()._process_event(watch_dir, event, self._get_new_file_archiver(watch_dir))
@@ -33,10 +29,7 @@ def owwatcher_object(monkeypatch):
     null_logger = logging.getLogger('owwatcher.null')
     null_logger.addHandler(logging.NullHandler())
 
-    null_syslog_logger = logging.getLogger('owwatcher.null-syslog')
-    null_syslog_logger.addHandler(logging.NullHandler)
-
-    return OWWatcherTest(monkeypatch, None, null_logger, null_syslog_logger)
+    return OWWatcherTest(monkeypatch, None, null_logger)
 
 def test_has_interesting_events_false(owwatcher_object):
     interesting_events = {iec.IN_ATTRIB, iec.IN_CREATE, iec.IN_MOVED_TO}
@@ -101,67 +94,71 @@ def test_is_world_writable_false(monkeypatch, owwatcher_object):
     assert not owwatcher_object._is_world_writable(path, filename)
 
 def test_process_event_no_interesting(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_OPEN, iec.IN_DELETE], "/tmp/random_dir_kljafl", "a_dir")
     patch_stat(monkeypatch, [0o777])
 
-    owwatcher_object._process_event("/tmp", event)
-    assert not owwatcher_object.syslog_logger.warning.called
+    owwatcher_object._process_event(watch_dir, event)
+    assert not owwatcher_object.alerter.add_event_to_alert_queue.called
 
 def test_process_event_ow_dir(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/random_dir_kljafl", "a_dir")
     patch_stat(monkeypatch, [0o002])
-    monkeypatch.setattr(os.path, "isdir", lambda _: True)
 
-    owwatcher_object._process_event("/tmp", event)
-    assert owwatcher_object.syslog_logger.warning.called
-    owwatcher_object.syslog_logger.warning.assert_called_with("Found world writable directory: /tmp/random_dir_kljafl/a_dir")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_ow_file(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/random_dir_kljafl", "a_file")
     patch_stat(monkeypatch, [0o002])
-    monkeypatch.setattr(os.path, "isdir", lambda _: False)
 
-    owwatcher_object._process_event("/tmp", event)
-    assert owwatcher_object.syslog_logger.warning.called
-    owwatcher_object.syslog_logger.warning.assert_called_with("Found world writable file: /tmp/random_dir_kljafl/a_file")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_perms_mask_file(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/random_dir_kljafl", "a_file")
     patch_stat(monkeypatch, [0o750])
-    monkeypatch.setattr(os.path, "isdir", lambda _: False)
 
     owwatcher_object.perms_mask = 0o010
-    owwatcher_object._process_event("/tmp", event)
-    assert owwatcher_object.syslog_logger.warning.called
-    owwatcher_object.syslog_logger.warning.assert_called_with("Found permissions matching mask 010 on file: /tmp/random_dir_kljafl/a_file")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_perms_mask_directory(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/random_dir_kljafl", "a_dir")
     patch_stat(monkeypatch, [0o750])
-    monkeypatch.setattr(os.path, "isdir", lambda _: True)
 
     owwatcher_object.perms_mask = 0o010
-    owwatcher_object._process_event("/tmp", event)
-    assert owwatcher_object.syslog_logger.warning.called
-    owwatcher_object.syslog_logger.warning.assert_called_with("Found permissions matching mask 010 on directory: /tmp/random_dir_kljafl/a_dir")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_mismatched_mask(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/random_dir_kljafl", "a_dir")
     patch_stat(monkeypatch, [0o722])
 
     owwatcher_object.perms_mask = 0o055
-    owwatcher_object._process_event("/tmp", event)
-    assert not owwatcher_object.syslog_logger.warning.called
+    owwatcher_object._process_event(watch_dir, event)
+    assert not owwatcher_object.alerter.add_event_to_alert_queue.called
 
 def test_process_event_empty_mask(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/random_dir_kljafl", "a_dir")
     patch_stat(monkeypatch, [0o777])
 
     owwatcher_object.perms_mask = 0o000
     owwatcher_object._process_event("/tmp/", event)
-    assert not owwatcher_object.syslog_logger.warning.called
+    assert not owwatcher_object.alerter.add_event_to_alert_queue.called
 
 def test_process_event_raise_fnf(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     def raise_(x):
         raise x
 
@@ -175,65 +172,58 @@ def test_process_event_raise_fnf(monkeypatch, owwatcher_object):
         assert False
 
 def test_process_event_ow_snap_path(monkeypatch, owwatcher_object):
-    event = (None, [iec.IN_CREATE, iec.IN_DELETE], "%s/tmp/dir1/dir2" % owwatcher.SNAP_HOSTFS_PATH_PREFIX, "a_file")
+    watch_dir = "%s/tmp" % owwatcher.SNAP_HOSTFS_PATH_PREFIX
+    event = (None, [iec.IN_CREATE, iec.IN_DELETE], "%s/dir1/dir2" % watch_dir, "a_file")
     patch_stat(monkeypatch, [0o002, 0o002])
-    monkeypatch.setattr(os.path, "isdir", lambda _: False)
 
     owwatcher_object.is_snap = True
-    owwatcher_object._process_event(owwatcher.SNAP_HOSTFS_PATH_PREFIX + "/tmp", event)
-    assert owwatcher_object.syslog_logger.warning.called
-    owwatcher_object.syslog_logger.warning.assert_called_with("Found world writable file: /tmp/dir1/dir2/a_file")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_ow_snap_dir(monkeypatch, owwatcher_object):
-    event = (None, [iec.IN_CREATE, iec.IN_DELETE], "%s/tmp" % owwatcher.SNAP_HOSTFS_PATH_PREFIX, "dir1")
+    watch_dir = "%s/tmp" % owwatcher.SNAP_HOSTFS_PATH_PREFIX
+    event = (None, [iec.IN_CREATE, iec.IN_DELETE], watch_dir, "dir1")
     patch_stat(monkeypatch, [0o002])
-    monkeypatch.setattr(os.path, "isdir", lambda _: True)
 
     owwatcher_object.is_snap = True
-    owwatcher_object._process_event(owwatcher.SNAP_HOSTFS_PATH_PREFIX + "/tmp", event)
-    assert owwatcher_object.syslog_logger.warning.called
-    owwatcher_object.syslog_logger.warning.assert_called_with("Found world writable directory: /tmp/dir1")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_directory_protects_ow_file(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/dir1/dir2", "a_file")
     patch_stat(monkeypatch, [0o702, 0o702, 0o700])
-    monkeypatch.setattr(os.path, "isdir", lambda _: False)
 
-    owwatcher_object._process_event("/tmp", event)
-    assert not owwatcher_object.syslog_logger.warning.called
-    assert owwatcher_object.syslog_logger.info.called
-    owwatcher_object.syslog_logger.info.assert_called_with("Found world " \
-            "writable file: /tmp/dir1/dir2/a_file -- Vulnerabilities are " \
-            "potentially mitigated as one or more parent directories do not " \
-            "have improperly configured permissions")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_directory_protects_ow_file_mask(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/dir1/dir2", "a_file")
     patch_stat(monkeypatch, [0o720, 0o720, 0o700])
-    monkeypatch.setattr(os.path, "isdir", lambda _: False)
 
     owwatcher_object.perms_mask = 0o077
-    owwatcher_object._process_event("/tmp", event)
-    assert not owwatcher_object.syslog_logger.warning.called
-    assert owwatcher_object.syslog_logger.info.called
-    owwatcher_object.syslog_logger.info.assert_called_with("Found permissions " \
-            "matching mask 077 on file: /tmp/dir1/dir2/a_file -- " \
-            "Vulnerabilities are potentially mitigated as one or more " \
-            "parent directories do not have improperly configured " \
-            "permissions")
+    owwatcher_object._process_event(watch_dir, event)
+    assert owwatcher_object.alerter.add_event_to_alert_queue.called
+    owwatcher_object.alerter.add_event_to_alert_queue.assert_called_with(watch_dir, event[2], event[3])
 
 def test_process_event_no_perms_mask_no_alert_no_archive_file(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CLOSE_WRITE], "/tmp/dir1/dir2", "a_file")
     patch_stat(monkeypatch, [0o700])
 
-    owwatcher_object._process_event("/tmp", event)
+    owwatcher_object._process_event(watch_dir, event)
     assert len(owwatcher_object.file_archivers) == 1
     assert not owwatcher_object.file_archivers[0].add_event_to_archive_file_queue.called
 
 def test_process_event_no_close_write_event(monkeypatch, owwatcher_object):
+    watch_dir = "/tmp"
     event = (None, [iec.IN_CREATE, iec.IN_DELETE], "/tmp/dir1/dir2", "a_file")
     patch_stat(monkeypatch, [0o777])
 
-    owwatcher_object._process_event("/tmp", event)
+    owwatcher_object._process_event(watch_dir, event)
     assert len(owwatcher_object.file_archivers) == 1
     assert owwatcher_object.file_archivers[0].add_event_to_archive_file_queue.called